requirements.txt
google-generativeai
imap-tools
pandas
python-dotenv